Complete implementation of port configuration
Rename the library query with item routines to have the type of the received
argument in the name as well. Perhaps the best answer long term would be to
receve a union as the argument and let the driver decide which value to use
out of the union rather than having numerous combinations of
query<type>with<type>item function calls which effectively all do the same
thing.