Fix bug in stall detection
If a motor stalls between the time it is asked to move and the wait command
(cli) is issued, the stall will not be detected. This will likely be the
case for cli when more than one motor is commanded. If two motors are asked
to move and then both the motors are waited on. The second will not be
waited upon until the first complete's is movement. If the second stalls
before the first motor completes its movement, the stall would not be
detected because the wait wasn't issued for the second motor yet.

This addresses the issue by always subscribing to the move event after a
move is issued. When "wait" is issued, the event will be waited on if it
isn't already set.