Fix possible race in motion completion callback
If another move is issued while the motor is being queried from completion
information in mdrive_completion_correct(), the completion corection
algorithm would reschedule the completion correct function because the motor
was still moving when checked. This fixes the race by detecting the callback
id before and after the chat with the device about motion completion
information. If the callback id is changed then the motion completion event
has already been cancelled.