Allow a disconnect from the motor in the CLI
Which will cleanly allow motors to be flash and disconnected from the board
and a different motor connected in manufacturing. The problem is that, if a
motor is connected to the daemon, firmware is checked and then flashed, the
motor is disconnected and put back on the shelf and a *different* motor is
connected, when the firmware is queried, the motor will not be asked again,
because the driver will have cached the firmware version, assuming that the
motor will not physically change over the duration of the running process.

mcDisconnect (<Python> Motor.disconnect) will allow the driver's destroy
method to be called at any time. Before, it could only be called at the exit
of the daemon.