Allow distinction between error return and string size
Change anonymous union in the (struct motor_query) to alleviate compiler
warnings, and embed a struct for PEEK/POKE strings where the buffer and the
buffer and the size of the buffer can be packed separately. This allows the
PEEK/POKE functions to return clear error codes and embed the returned
string size in the motor_query handled.