Remove (more) dependence on readahead
Wait less time for extra characters to roll in, when reading from the
device's fd. This was originally implemented to compensate for the code to
detect chars after the '\n' char in the response. Since the code reads ahead
of the '\n' char, more time was needed to wait to capture the following
characters. The code is slowly being rewritten to perform read-behinds now
to react to previous characters rather than following characters. Also, the
extra wait time seems to screw up firmware upgrade mode responses in some
cases. Overall, waiting less will slightly increase performance when reading
single ACKs off the line, which will be the case most of the time.

Somehow, the motors will echo a '\r' CR character with an '\x13' character.
While probably an IMS bug, quietly accept it as an echo response terminator
if followed by an ascii ACK ('\x06').