(Almost) keep track of default assigns vs. assigns in code
Defaults should be collected and outputted separately from the rest of the
code. Otherwise, modules that are included and expect various defaults to be
set will not be compiled correctly if their defaults are written in the
middle of code with labels and such. Technically, defaults should be sent
before program mode is entered or after its exit