First successful mockup of formal Proxy/Stub model
Allows for far less magic between the client and server pieces of the
mcontrol library calls. In-process clients can now call the Impl functions
without an extra trampoline involving packing arguments up into the message
structure used by the out-of-process clients. Best of all, the Impl
functions can now use standard 'return' statements and consistent
marshalling. In other words, the out-of-process return and parameter
marshalling is handled in the Proxy and Stub functions. And the parameter
list for the Impl function looks the same as for the client-side function.

One layer of magic remains: the CONTEXT magic parameter. The caller of the
Impl function will construct and pass a CONTEXT parameter which will provide
information about the caller (pid, outofprocess, and an inprocess). Also, if
the function declares to receive a MOTOR argument, the client-side motor id
handle is checked and converted to the server-side (Motor *). Checks are
performed before the Impl function is called, so the Impl function can
safely assume that (CONTEXT->motor) is a valid pointer to a (Motor *).