Fix slight disagreement
in the location of the motor id number. I'd like to move away from a special
member of the request_message structure and leave the motor id number in the
args structure specific to each call. This method more closely maps to the
actual calling convention and doesn't make the assumption that every call is
made with respect to a motor (like tracing for instance).