Properly implement `fail` and friends in sub-scripts
The return statement should just signal the execute_script() function that
the script should be abandoned. The `call` command will then pop the call
stack. This will allow things like `fail` to properly affect the call stack
as well. Lastly, the call statement should signal the execute_script()
function at the current level to abort if the script state is not `RUNNING`