Add block support to cli scripts
in the form of do ... done. This will assist more complex if statements and
will allow for formal function definitions and loops.

The `if` statement was reworked along with the parser to split commands by
semi-colons and colons (backwards compatibility). The if statement will now
skip the command following if the condition does not evaluate to a boolean
true.

Nested blocks should be handled transparently, so ifs inside of ifs should
be handled correctly