cli: fix erroneous error message at test end
If the test script does not complete with one of 'succeed', 'fail', or
'abort', the code used to implement do/done blocks added a 'return'
statement ahead of the blocks. If the return statement was actually executed
because the script did not use one of the other exit techniques, an error
would be given since the stack was unbalanced.

The environment would have assumed success if the test is completed without
any exit given, so it makes the most sense to end with a forced 'succeed'
statement.