microcode: Fix incorrect HT and MT settings
The original microcode included HT=0 and MT=0, HT being the worse of the
two. With HT=0, the hold current of the MDrive motors is never engaged.
Instead, the current on the motor coils will remain at the run current
indefinitely. The MT setting is recommended for motor settling time, again,
to be a non-zero value