cli: test: mockup tesk tasklets
Tasklets allow for parallel testing, where part of the test can be run in
parallel with another part of the test. Tasklets work with something like
Python's generator/coroutine functions where the tasklets can be
communicated with using `yield` inside the tasklet and `tasklet <> read` to
retrieve data from the tasklet, and `tasklet <> send {x}` and
`let x = [yield]` inside the tasklet.

Crashes in the main test will propogate to the inner tests, aborting them,
and finishing status other than success will be propogated to the main test
with `tasklet <> join`

More to follow in the documentation