cli: test: fix race on send/yield
Previously, it was possible for send to be released and called again before
the do_yield() method would consume the value of inkitty. This patch fixes
the issue by forcing send() to wait for the task to reach a yield command
before it delivers the value sent. Doing so creates more of a handshake
rather than a handoff