cli: test: rollup task completion status
When a test completes sub-tasks, allow the completion status of a task to
propagate to other tasks as well as the parent test. Since statuses are
already increasing numerically with respect to significance, this algorithm
just rolls up all the status of the parent test and sub tasks so that the
highest completion status will win out.