mdrive: Allow more time for the delayed response
The 0.1 version allowed up to 70ms for the early ACK and the delayed
response. This new version will drop the initial timeout when the early ACK
is received and only wait the {more_waittime} for the following delayed
response from the device. 25ms seems to catch over 99.99% of the delayed
responses; however it seems to need to be increased to allow from some edge
cases.

Also, if the early ACK was received and the timeout of the delayed response
is too short and no (complete) response is received, the transmission should
be retried. As written, the response was reclassified, forgetting the
timeout already marked. Since a standalone ACK is considered a successful
transmission by the classification algorithm, the transmission would not
have been retried even though data was expected but not received. This patch
will only classify the response if not already marked TIMEOUT